Main: do not truncate uint16_t to uint8_t in usbsysfsReadDevicePropertyU16Def()
svn:sync-xref-src-repo-rev: r108887

+++ src/VBox/Main/src-server/linux/USBGetDevices.cpp
@@ -1,4 +1,4 @@
-/* $Id: USBGetDevices.cpp 60373 2016-04-07 14:21:30Z alexander.eichner@oracle.com $ */
+/* $Id: USBGetDevices.cpp 62343 2016-07-20 09:02:26Z noreply@oracle.com $ */
 /** @file
  * VirtualBox Linux host USB device enumeration.
  */
@@ -1261,7 +1261,7 @@ static uint8_t usbsysfsReadDevicePropertyU8Def(unsigned uBase, uint8_t bDef, con
  * @param   pszFormat   The format string for the property.
  * @param   ...         Arguments for the format string.
  */
-static uint8_t usbsysfsReadDevicePropertyU16Def(unsigned uBase, uint16_t u16Def, const char *pszFormat, ...)
+static uint16_t usbsysfsReadDevicePropertyU16Def(unsigned uBase, uint16_t u16Def, const char *pszFormat, ...)
 {
     int64_t i64Tmp = 0;
